Move sessions-related tests out of UsersControllerTest
